Consolidate chaos UGens
